requirements.txt
SQLAlchemy==2.0.38
python-dotenv==1.0.1